add PageLayout, header and footer components
