+++ apps/ui/src/app/[lang]/second-page/Second.tsx
@@ -24,10 +24,10 @@ import { getTranslation, ValidLanguage } from '@/i18n';
 const Second = async ({ lang }: { lang: ValidLanguage }) => {
   const translate = await getTranslation(lang, 'second-page');
   return (
-    <main className="flex min-h-screen flex-col items-center justify-between p-24">
+    <div>
       <h1>{translate('title')}</h1>
       <Link href={`/${lang}`}>{translate('back-to-home')}</Link>
-    </main>
+    </div>
   );
 };
