+++ apps/ui/src/app/[lang]/second-page/page.tsx
@@ -17,7 +17,6 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import Header from '@/components/Header';
 import Second from './Second';
 import { ValidLanguage } from '@/i18n';
 
@@ -26,10 +25,5 @@ export default async function Page({
 }: {
   params: { lang: ValidLanguage };
 }) {
-  return (
-    <>
-      <Header lang={lang} />
-      <Second lang={lang} />;
-    </>
-  );
+  return <Second lang={lang} />;
 }
