+++ apps/ui/src/components/Container/Container.module.scss
@@ -1,0 +1,3 @@
+.container {
+    height: inherit;
+}
