+++ apps/ui/src/components/Footer/Footer.module.scss
@@ -1,0 +1,6 @@
+.footer {
+    background-color: theme('colors.grayscale.700');
+    color: theme('colors.grayscale.100');
+    height: 162px;
+    padding: 0.5rem;
+}
