+++ apps/ui/src/components/Header/Header.module.scss
@@ -1,0 +1,9 @@
+.header {
+    height: 80px;
+    background-color: theme('colors.grayscale.100');
+    padding: 0.5rem;
+}
+
+.logo {
+
+}
