+++ apps/ui/src/components/Header/index.tsx
@@ -16,42 +16,31 @@
  * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-import { Montserrat } from 'next/font/google';
 
-import '../globals.css';
-
-import { ValidLanguage } from '@/i18n';
+import { ValidLanguage, getTranslation } from '@/i18n';
 import { supportedLanguages } from '@/i18n/settings';
-import PageLayout from '@/components/PageLayout';
 
-const montserrat = Montserrat({
-  subsets: ['latin'],
-  display: 'swap',
-  variable: '--font-montserrat',
-});
+import LanguageToggle from '@/components/LanguageToggle';
+import styles from './Header.module.scss';
 
-export async function generateStaticParams() {
-  return supportedLanguages.map((lang) => ({ lang }));
-}
-
-// TODO: translate metadata
-export const metadata = {
-  title: 'OHCRN - Homepage',
-  description: 'Landing page for OHCRN Patient Enrolment Portal',
+export const getUnselectedLang = (lang: ValidLanguage): string => {
+  return supportedLanguages.filter((l) => l !== lang)[0];
 };
 
-export default async function RootLayout({
-  children,
-  params: { lang },
-}: {
-  children: React.ReactNode;
-  params: { lang: ValidLanguage };
-}) {
+const Header = async ({ lang }: { lang: ValidLanguage }) => {
+  const translate = await getTranslation(lang, 'header');
+  const langToSelect = getUnselectedLang(lang);
+
   return (
-    <html lang={lang}>
-      <body className={`${montserrat.className}`}>
-        <PageLayout lang={lang}>{children}</PageLayout>
-      </body>
-    </html>
+    <header className={styles.header}>
+      <LanguageToggle
+        displayLangToSelect={translate(langToSelect)}
+        selected={lang}
+      >
+        {translate(lang)}
+      </LanguageToggle>
+    </header>
   );
-}
+};
+
+export default Header;
