+++ apps/ui/src/components/PageLayout/PageLayout.module.scss
@@ -1,0 +1,11 @@
+.main {
+    background-color: theme('colors.grayscale.200');
+    height: inherit;
+    padding: 1rem 3rem;
+    display: flex;
+    justify-content: center;
+}
+
+.content {
+    background-color: theme('colors.grayscale.100');
+}
