+++ apps/ui/tailwind.config.js
@@ -17,11 +17,7 @@ module.exports = {
       // screen size breakpoints
     },
     extend: {
-      backgroundImage: {
-        'gradient-radial': 'radial-gradient(var(--tw-gradient-stops))',
-        'gradient-conic':
-          'conic-gradient(from 180deg at 50% 50%, var(--tw-gradient-stops))',
-      },
+
     },
   },
   plugins: [],
